api/requirements/base.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.40
alembic==1.15.2
psycopg[binary]==3.2.13
PyJWT==2.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
